# Optimized the distance computations