--- a/debian/changelog
+++ b/debian/changelog
@@ -1,3 +1,7 @@
+  [Michele Simionato]
+  * Removed the speedups and optimized the distance calculations at the
+    numpy level
+
 python-oq-hazardlib (0.21.0-0~precise01) precise; urgency=low
 
   [Trevor Allen/Graeme Weatherill]