--- a/openquake/hazardlib/geo/geodetic.py
+++ b/openquake/hazardlib/geo/geodetic.py
@@ -239,7 +239,7 @@ def min_geodetic_distance(mlons, mlats, slons, slats, diameter=2*EARTH_RADIUS):
     mlons, mlats, slons, slats = _prepare_coords(mlons, mlats, slons, slats)
     cos_mlats = numpy.cos(mlats)
     cos_slats = numpy.cos(slats)
-    if len(mlons.shape) == 1:
+    if len(mlons.shape) == 1:  # fast lane
         result = numpy.zeros((len(mlons), len(slons)))
         for j in range(len(mlons)):
             a = numpy.sin((mlats[j] - slats) / 2.0)