… openquake/hazardlib/geo/mesh.py
… openquake/hazardlib/geo/polygon.py
… openquake/hazardlib/geo/surface/base.py
… openquake/hazardlib/source/rupture.py
… openquake/hazardlib/tests/geo/geodetic_test.py
… openquake/hazardlib/tests/gsim/chiou_youngs_2014_test.py