Having removed the speedups, it is easy to optimize the distance computations so that we are faster than before while remaining pure Python. The trick is on the function `pure_distances`: there is an `if len(mlons) < len(slons)` that makes sure that the short loop in made in Python and the long loop in numpy (before it was the opposite).

Here are some numbers for a computation dominated by `make_context` (a reduces classical PSHA USA model with 1000 sites) which is representative of the worst case scenario:

```
master
making contexts  1,041 s
total runtime 275s

no-speedups
making contexts  1948 s
total runtime 334s

no-speedups-2
making contexts  1,032 s
total runtime 256s
```

Most large calculations are dominated by the PoEs computation, not by `making context`, so the distance computations make little difference (this is even more true for disaggregation).