@larsbutler 2016-10-15
@micheles Cool patch, Michele! I'm curious to know what array sizes you tested with. If memory serves, arrays of up to 10^4 elements did not gain anything from the C ext speedups (and might have been slower, even), but above that the benefit was something like 14x.