@micheles 2016-10-17
Hi Lars, good to hear from you!
I tested with a variety of array sizes, but the common case in hazard calculations is for array of 1000-2000 sites. The root of the problem is clear when one consider the distance between a planar rupture (4 vertices) and a site collection of say 1000 sites. Computing 4 distances for each site 1000 times is inefficient (1000 arrays of 4 elements) and I saw an order of magnitude speedup by going to C. However, by changing the loop and computing 1000 distances 4 times (4 arrays of 1000 elements), there is no difference between numpy and C, or numpy can be even slightly more efficient. This is why I could remove the speedups without losing any performance.