@larsbutler 2016-10-17
@micheles That's awesome. Thanks for the explanation. Nice work!